LF-4746 Convert hook to TS and create types file shared between component and hook
