+++ packages/webapp/src/components/Finances/DateRangeSelector/useFinancesDateRange.ts
@@ -12,20 +12,28 @@
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
  */
+import { Moment } from 'moment';
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import { dateRangeDataSelector } from '../../../containers/Finances/selectors';
+import { setDateRange } from '../../../containers/Finances/actions';
 import DateRange, { MONDAY, SUNDAY } from '../../../util/dateRange';
 import { DateRangeOptions } from '../../DateRangeSelector/constants';
-import { setDateRange } from '../../../containers/Finances/actions';
+import { DateRangeSelection } from './types';
+
+interface UseFinancesDateRangeProps {
+  weekStartDate?: typeof SUNDAY | typeof MONDAY;
+}
 
 /**
- * Returns startDate and endDate in the store.
- * If they are not defined, set the default option and dates.
+ * Returns startDate and endDate from the store.
+ * If they are not defined, sets the default option and dates.
  */
-export default function useFinancesDateRange({ weekStartDate = SUNDAY }) {
-  const dateRange = useSelector(dateRangeDataSelector);
+export default function useFinancesDateRange({
+  weekStartDate = SUNDAY,
+}: UseFinancesDateRangeProps): { startDate?: string | Moment; endDate?: string | Moment } {
+  const dateRange: DateRangeSelection = useSelector(dateRangeDataSelector);
   const dispatch = useDispatch();
 
   const dateRangeUtil = new DateRange(new Date(), weekStartDate);
